requirements.txt
ultralytics
streamlit
pandas
pillow
watchdog